requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
deprecation==2.1.0
Django ==  4.1
django-cors-headers==4.3.1
django-dotenv==1.4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
et-xmlfile == 1.1.0
gotrue==2.1.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
numpy>=1.24.0,<=1.26.1
openpyxl == 3.0.10
packaging==23.2
pandas==2.1.4
postgrest==0.15.0
psycopg2-binary==2.9.9
pydantic==2.6.1
pydantic_core==2.16.2
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
realtime==1.0.2
requests==2.31.0
setuptools==68.0.0
six==1.16.0
sniffio==1.3.0
sqlparse == 0.4.4
storage3==0.7.0
StrEnum==0.4.15
supabase==2.3.4
supafunc==0.3.3
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
websockets==11.0.3
wheel==0.41.2
celery==5.3.6
redis >=5.0.2
gunicorn >= 21.2.0



